Move common behavior from keyboards into userspace
