+++ keyboards/kbd6x/keymaps/konstantin/config.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#define LAYER_FN
