+++ keyboards/kbd6x/keymaps/konstantin/keymap.c
@@ -1,100 +1,5 @@
 #include QMK_KEYBOARD_H
-
-#define TOP     LCTL(KC_HOME)
-#define BOTTOM  LCTL(KC_END)
-#define DSKTP_L LCTL(LGUI(KC_LEFT))
-#define DSKTP_R LCTL(LGUI(KC_RGHT))
-
-#define FN      MO(L_FN)
-#define FN_FNLK TT(L_FN)
-#define LCT_CPS LCTL_T(KC_CAPS)
-
-#define DESKTOP TD(TD_DESKTOP)
-#define LSFT_FN TD(TD_LSFT_FN)
-#define RAL_RGU TD(TD_RAL_RGU)
-
-enum layers {
-  L_BASE,
-  L_FN,
-};
-
-enum custom_keycodes {
-  CLEAR = SAFE_RANGE,
-};
-
-bool process_record_user(uint16_t keycode, keyrecord_t *record) {
-  switch (keycode) {
-  case CLEAR:
-    if (record->event.pressed) {
-      SEND_STRING(SS_LCTRL("a") SS_TAP(X_DELETE));
-    }
-    return false;
-
-  default:
-    return true;
-  }
-}
-
-void td_lsft_fn_each(qk_tap_dance_state_t *state, void *user_data) {
-  // Single tap → LShift, double tap → Fn, triple tap etc. → Fn+LShift
-  if (state->count == 1 || state->count == 3) {
-    register_code(KC_LSFT);
-  } else if (state->count == 2) {
-    unregister_code(KC_LSFT);
-    // Prevent tap dance from sending LShift as a weak mod
-    state->weak_mods &= ~MOD_BIT(KC_LSFT);
-    layer_on(L_FN);
-  }
-}
-
-void td_lsft_fn_reset(qk_tap_dance_state_t *state, void *user_data) {
-  if (state->count == 1 || state->count >= 3) {
-    unregister_code(KC_LSFT);
-  }
-  if (state->count >= 2) {
-    layer_off(L_FN);
-  }
-}
-
-#define ACTION_TAP_DANCE_DOUBLE_MODS(mod1, mod2) { \
-    .fn = { td_double_mods_each, NULL, td_double_mods_reset }, \
-    .user_data = &(qk_tap_dance_pair_t){ mod1, mod2 }, \
-  }
-
-void td_double_mods_each(qk_tap_dance_state_t *state, void *user_data) {
-  qk_tap_dance_pair_t *mods = (qk_tap_dance_pair_t *)user_data;
-  // Single tap → mod1, double tap → mod2, triple tap etc. → mod1+mod2
-  if (state->count == 1 || state->count == 3) {
-    register_code(mods->kc1);
-  } else if (state->count == 2) {
-    unregister_code(mods->kc1);
-    register_code(mods->kc2);
-  }
-  // Prevent tap dance from sending kc1 and kc2 as weak mods
-  state->weak_mods &= ~(MOD_BIT(mods->kc1) | MOD_BIT(mods->kc2));
-}
-
-void td_double_mods_reset(qk_tap_dance_state_t *state, void *user_data) {
-  qk_tap_dance_pair_t *mods = (qk_tap_dance_pair_t *)user_data;
-  if (state->count == 1 || state->count >= 3) {
-    unregister_code(mods->kc1);
-  }
-  if (state->count >= 2) {
-    unregister_code(mods->kc2);
-  }
-}
-
-enum tap_dance {
-  TD_DESKTOP,
-  TD_LSFT_FN,
-  TD_RAL_RGU,
-};
-
-qk_tap_dance_action_t tap_dance_actions[] = {
-  [TD_DESKTOP] = ACTION_TAP_DANCE_DOUBLE(LCTL(LGUI(KC_D)), LCTL(LGUI(KC_F4))), // Add/close virtual desktop
-  [TD_LSFT_FN] = ACTION_TAP_DANCE_FN_ADVANCED(td_lsft_fn_each, NULL, td_lsft_fn_reset),
-  [TD_RAL_RGU] = ACTION_TAP_DANCE_DOUBLE_MODS(KC_RALT, KC_RGUI),
-};
+#include "konstantin.h"
 
 const uint16_t PROGMEM keymaps[][MATRIX_ROWS][MATRIX_COLS] = {
   /* Base layer
