+++ users/konstantin/rules.mk
@@ -1,7 +1,3 @@
-SRC += konstantin.c
-
-EXTRAFLAGS += -flto
-
 BOOTMAGIC_ENABLE = full # TODO: Disable once Command is fixed
 COMMAND_ENABLE   = no
 CONSOLE_ENABLE   = yes
@@ -10,3 +6,10 @@ MOUSEKEY_ENABLE  = yes
 NKRO_ENABLE      = yes
 TAP_DANCE_ENABLE = yes
 UNICODE_ENABLE   = yes
+
+SRC += konstantin.c
+ifeq ($(strip $(TAP_DANCE_ENABLE)), yes)
+  SRC += tap_dance.c
+endif
+
+EXTRAFLAGS += -flto
