+++ users/konstantin/unicode.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "quantum.h"
+
+#define COMMA   UC(0x002C)
+#define L_PAREN UC(0x0028)
+#define R_PAREN UC(0x0029)
+#define EQUALS  UC(0x003D)
+#define TIMES   UC(0x00D7)
+#define DIVIDE  UC(0x00F7)
+#define MINUS   UC(0x2212)
